+++ src/pacman/Ghost.java
@@ -1,8 +1,7 @@
 package pacman;
 
 public class Ghost extends Character{
-
-    Ghost(String characterIconPath, int sizeX, int sizeY) {
+    Ghost(String characterIconPath, float sizeX, float sizeY) {
         super(characterIconPath, sizeX, sizeY);
     }
 }
