+++ src/pacman/Player.java
@@ -1,7 +1,7 @@
 package pacman;
 
 public class Player extends Character {
-    Player(String iconImage, int sizeX, int sizeY) {
+    Player(String iconImage, float sizeX, float sizeY) {
         super(iconImage,sizeX,sizeY);
     }
 }
